feat: display scaffold messenger if not enough funds to bind
