+++ lib/components/account_box.dart
@@ -11,6 +11,7 @@ import 'package:reef_mobile_app/components/modals/export_qr_account_modal.dart';
 import 'package:reef_mobile_app/model/ReefAppState.dart';
 import 'package:reef_mobile_app/model/account/ReefAccount.dart';
 import 'package:reef_mobile_app/model/status-data-object/StatusDataObject.dart';
+import 'package:reef_mobile_app/utils/bind_evm.dart';
 import 'package:reef_mobile_app/utils/functions.dart';
 
 import '../utils/styles.dart';
@@ -186,6 +187,7 @@ class _AccountBoxState extends State<AccountBox> {
       StatusDataObject<ReefAccount> reefAccount, bool lightTheme) {
     var textColor1 = lightTheme ? Styles.textColor : Colors.white;
     var textColor2 = lightTheme ? Colors.black38 : Styles.textLightColor;
+
     return Flex(
         direction: Axis.vertical,
         crossAxisAlignment: CrossAxisAlignment.center,
@@ -273,6 +275,22 @@ class _AccountBoxState extends State<AccountBox> {
                         borderRadius: BorderRadius.circular(12)),
                     child: TextButton(
                         onPressed: () {
+                          var availableAccounts = getSignersWithEnoughBalance(
+                              widget.reefAccountFDM.data);
+                          print("anuna $availableAccounts");
+                          var hasBalance =
+                              hasBalanceForBinding(widget.reefAccountFDM.data);
+
+                          if (availableAccounts.isEmpty && !hasBalance) {
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              SnackBar(
+                                content: Text(AppLocalizations.of(context)!
+                                    .account_imported_successfully),
+                                duration: Duration(seconds: 2),
+                              ),
+                            );
+                            return;
+                          }
                           showBindEvmModal(context,
                               bindFor: widget.reefAccountFDM.data);
                         },
