+++ lib/l10n/app_hi.arb
@@ -152,5 +152,6 @@
   "completed_binance_purchase":"यदि आपने बिनेंस कनेक्ट पर अपनी खरीदारी पूरी कर ली है, तो आपके रीफ बैलेंस को चेन पर लेन-देन की पुष्टि होने के बाद अपडेट किया जाएगा।",
   "go_to_binance":"बिनेंस कनेक्ट पर जाएं",
   "biometric_auth":"बायोमेट्रिक प्रमाणीकरण",
-  "dont_send_funds_here":"यहां पैसे मत भेजो"
+  "dont_send_funds_here":"यहां पैसे मत भेजो",
+  "account_imported_successfully":"अपर्याप्त धन! कृपया अपना रीफ वॉलेट भरें"
 }
